refactor: ReactElement return type 설정
