+++ src/components/auth/LoginForm.tsx
@@ -1,4 +1,4 @@
-import { Button } from "components/common/Button";
+import Button from "components/common/Button";
 import Input from "components/common/Input";
 import React, { ReactElement } from "react";
 import styled from "styled-components";
